ulog: klog only on Linux.
